+++ src/sonos-get-status.js
@@ -4,11 +4,10 @@ var helper = new SonosHelper();
 module.exports = function (RED) {
   'use strict';
 
+  /**  Create Get Status Node and subscribe to messages
+  * @param  {object} config current node configuration data
+  */
   function SonosGetStatusNode (config) {
-    /**  Create Get Status Node and subscribe to messages
-    * @param  {object} config current node configuration data
-    */
-
     RED.nodes.createNode(this, config);
 
     // verify config node. if valid then set status and subscribe to messages
@@ -64,177 +63,347 @@ module.exports = function (RED) {
       node.error('validate payload - invalid payload. Details' + JSON.stringify(msg.payload));
       return;
     }
-
     let command = String(msg.payload);
     command = command.toLowerCase();
 
     // dispatch
-    if (command === 'get_stateonly') {
-      getSonosCurrentState(node, msg, sonosPlayer, false);
-    } else if (command === 'get_basics') {
-      getSonosCurrentState(node, msg, sonosPlayer, true);
+    if (command === 'get_basics') {
+      // CAUTION last parameter initiates a "chain" --> async
+      getPlayerStateV2(node, msg, sonosPlayer, 'get_basics');
+    } else if (command === 'get_stateonly') {
+      // for compatibility reasons - depreciated
+      getPlayerStateV2(node, msg, sonosPlayer, 'separate');
+    } else if (command === 'get_state') {
+      getPlayerStateV2(node, msg, sonosPlayer, '');
+    } else if (command === 'get_volume') {
+      getPlayerVolumeV2(node, msg, sonosPlayer, false);
+    } else if (command === 'get_muted') {
+      getPlayerMutedV2(node, msg, sonosPlayer, false);
+    } else if (command === 'get_sonosname') {
+    // for compatibility reasons - depreciated
+      getPlayerNameV2(node, msg, sonosPlayer, 'separate');
+    } else if (command === 'get_name') {
+      getPlayerNameV2(node, msg, sonosPlayer, '');
+    } else if (command === 'get_properties') {
+      getPlayerProperties(node, msg, sonosPlayer);
     } else if (command === 'get_songmedia') {
-      getSonosCurrentSong(node, msg, sonosPlayer, true);
+      // CAUTION last parameter initiates a "chain" --> async
+      getPlayerCurrentSong(node, msg, sonosPlayer, 'get_songmedia');
     } else if (command === 'get_songinfo') {
-      getSonosCurrentSong(node, msg, sonosPlayer, false);
+      getPlayerCurrentSong(node, msg, sonosPlayer, '');
     } else if (command === 'get_mediainfo') {
-      getSonosMediaInfo(node, msg, sonosPlayer, false);
+      getMediaInfo(node, msg, sonosPlayer, '');
     } else if (command === 'get_positioninfo') {
-      getSonosPositionInfo(node, msg, sonosPlayer, false);
+      getPositionInfo(node, msg, sonosPlayer, '');
     } else {
       node.status({ fill: 'green', shape: 'dot', text: 'warning:depatching commands - invalid command' });
       node.warn('depatching commands - invalid command: ' + command);
     }
   }
 
-  // ------------------------------------------------------------------------------------------
+  // -----------------------------------------------------
+  // start chain get_basics
+  // -----------------------------------------------------
 
-  /**  Validate sonos player and input message then get state and all other data.
+  /** Gets the sonos player state and either outputs or starts another asyn request (chaining).
   * @param  {Object} node current node
   * @param  {Object} msg incoming message
-  * @param  {Object} sonosPlayer SONOS player object
-  * @param  {Boolean} chain start request for other status information (chaining)
-  * changes msg.state
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.state or msg.payload
   */
-  function getSonosCurrentState (node, msg, sonosPlayer, chain) {
-    // execute first api to get state
-    const sonosFunction = 'player state';
-    let errorShort = 'invalid player state received';
+  function getPlayerStateV2 (node, msg, sonosPlayer, outputType) {
+    const sonosFunction = 'get player state';
+    let msgShort = '';
     sonosPlayer.getCurrentState().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid player state received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
-      msg.state = response;
-      if (chain) {
-        node.debug('got valid player state');
-        getSonosVolume(node, msg, sonosPlayer, true);
-      } else {
-        node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
-        node.debug(`ok:${sonosFunction}`);
-        node.send(msg);
+
+      // output chaining, to seperate msg property or to msg.payload
+      node.debug('got valid player state');
+      switch (outputType) {
+        case 'get_basics':
+          // part of a chain, dont send message
+          msg.state = response;
+          node.debug('chain next command');
+          getPlayerVolumeV2(node, msg, sonosPlayer, 'get_basics');
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.state = response;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.debug('output to payload');
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
       }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  /**  get sonos volume for selected player and continue in chain.
+  /** Gets the sonos player volume and either outputs or starts another asyn request (chaining).
   * @param  {Object} node current node
   * @param  {Object} msg incoming message
-  * @param  {Object} sonosPlayer SONOS player object
-  * @param  {Boolean} chain start request for other status information (chaining)
-  * changes msg.volume, msg.normalized_volume
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.volume and msg.normalized_volume or msg.payload
   */
-  function getSonosVolume (node, msg, sonosPlayer, chain) {
-    const sonosFunction = 'volume';
-    let errorShort = 'invalid volume received';
+  function getPlayerVolumeV2 (node, msg, sonosPlayer, outputType) {
+    const sonosFunction = 'get player volume';
+    let msgShort = '';
     sonosPlayer.getVolume().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid player volume received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
+
       if (response < 0 || response > 100) {
-        errorShort = 'invalid volume rage received';
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction}  - ${errorShort} Details: ` + JSON.stringify(response));
+        msgShort = 'invalid volume rage received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction}  - ${msgShort} Details: ` + JSON.stringify(response));
         return;
       }
-      // Output data
-      msg.volume = response;
-      msg.normalized_volume = response / 100.0;
-      node.debug('got valid volume');
-      getSonosMuted(node, msg, sonosPlayer, chain);
+
+      // output chaining, to seperate msg property or to msg.payload
+      node.debug('got valid player volume');
+      switch (outputType) {
+        case 'get_basics':
+          // part of a chain, dont send message
+          msg.volume = response;
+          msg.normalized_volume = response / 100.0;
+          node.debug('got volume data and continue');
+          getPlayerMutedV2(node, msg, sonosPlayer, 'get_basics');
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.volume = response;
+          msg.normalized_volume = response / 100.0;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
+      }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  function getSonosMuted (node, msg, sonosPlayer, chain) {
-    //   changes msg.muted
-    const sonosFunction = 'muted';
-    let errorShort = 'invalid muted received';
+  /** Gets the sonos player muted state and either outputs or starts another asyn request (chaining).
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.muted or msg.payload
+  */
+  function getPlayerMutedV2 (node, msg, sonosPlayer, outputType) {
+    const sonosFunction = 'get player muted state';
+    let msgShort = '';
     sonosPlayer.getMuted().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid muted state received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
-      // Output data
-      node.debug('got valid mute value');
-      msg.muted = response;
-      getSonosName(node, msg, sonosPlayer, chain);
+
+      // output chaining, to seperate msg property or to msg.payload
+      node.debug('got valid player muted state');
+      switch (outputType) {
+        case 'get_basics':
+          // part of a chain, dont send message
+          msg.muted = response;
+          getPlayerNameV2(node, msg, sonosPlayer, 'get_basics');
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.muted = response;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
+      }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  function getSonosName (node, msg, sonosPlayer, chain) {
-    //   changes msg.sonosName
-    const sonosFunction = 'SONOS name';
-    let errorShort = 'invalid SONOS name received';
+  /** Gets the sonos player name and either outputs or starts another asyn request (chaining).
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.sonosName or msg.payload
+  */
+  function getPlayerNameV2 (node, msg, sonosPlayer, outputType) {
+    const sonosFunction = 'get player name';
+    let msgShort = '';
     sonosPlayer.getName().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid player name received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
-      // Output data
-      node.debug('got valid Sonos player name');
-      msg.sonosName = response;
-      getSonosGroupAttributes(node, msg, sonosPlayer, chain);
+
+      // output chaining, to seperate msg property or to msg.payload
+      node.debug('got valid player name');
+      switch (outputType) {
+        case 'get_basics':
+          // part of a chain, dont send message
+          msg.sonosName = response;
+          getPlayerGroupAttributesV2(node, msg, sonosPlayer, 'get_basics');
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.sonosName = response;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
+      }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  function getSonosGroupAttributes (node, msg, sonosPlayer, chain) {
-    //   changes msg.sonosGroup
-    const sonosFunction = 'group attributes';
-    let errorShort = 'invalid group attributes received';
+  /** Gets the sonos player group attributes and either outputs or starts another asyn request (chaining).
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.sonosGroup or msg.payload
+  */
+  function getPlayerGroupAttributesV2 (node, msg, sonosPlayer, outputType) {
+    const sonosFunction = 'get group attributes';
+    let msgShort = '';
     sonosPlayer.zoneGroupTopologyService().GetZoneGroupAttributes().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid groupt attributes received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
-      // Output data
-      node.debug('got valid Groups');
-      msg.sonosGroup = response;
-      node.status({ fill: 'green', shape: 'dot', text: 'ok:get basics' });
-      node.debug('ok:get basics');
-      node.send(msg);
+
+      // output chaining, to seperate msg property or to msg.payload
+      node.debug('got group attributes');
+      switch (outputType) {
+        case 'get_basics':
+          // end of chain
+          msg.sonosGroup = response;
+          node.status({ fill: 'green', shape: 'dot', text: 'ok:get basics' });
+          node.debug('ok:get basics');
+          node.send(msg);
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.sonosGroup = response;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
+      }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  // next chain command for song, media, position
+  // -----------------------------------------
+  // end of chain (get basics)
+  // -----------------------------------------
 
-  function getSonosCurrentSong (node, msg, sonosPlayer, chain) {
-    // finally changes  msg.title (string), msg.artist (string), msg.song (obj) msg.media (obj), msg.position (obj)
-    // first step artist, title
+  /** Gets the sonos player properties and outputs to payload.
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * changes msg.payload as object
+  */
+  function getPlayerProperties (node, msg, sonosPlayer) {
+    const sonosFunction = 'get player properties';
+    let msgShort = '';
+    sonosPlayer.deviceDescription().then(response => {
+      if (response === null || response === undefined) {
+        msgShort = 'invalid player properties received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
+        return;
+      }
+
+      // Output data to payload
+      node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+      node.debug(`ok:${sonosFunction}`);
+      msg.payload = response;
+      node.send(msg);
+    }).catch(err => {
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
+    });
+  }
+
+  // ---------------------------------------------------
+  // start of chain get_songmedia
+  // ---------------------------------------------------
+
+  /** Gets the sonos player current song and either outputs or starts another asyn request (chaining).
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.artist, title, albumArt, song, media and position info in chain mode
+  */
+  function getPlayerCurrentSong (node, msg, sonosPlayer, outputType) {
     let artist = 'unknown';
     let title = 'unknown';
     let albumArtURL = 'unknown';
-    const sonosFunction = 'current song';
-    let errorShort = 'invalid current song received';
+    const sonosFunction = 'get current song';
+    let msgShort = '';
     sonosPlayer.currentTrack().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid current song received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
       // message albumArtURL property
@@ -256,41 +425,59 @@ module.exports = function (RED) {
           artist = response.title.split(' - ')[0];
           title = response.title.split(' - ')[1];
         } else {
-          errorShort = 'invalid combination artist title received';
-          node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-          node.error(`${sonosFunction}  - ${errorShort} Details: ` + JSON.stringify(response));
+          msgShort = 'invalid combination artist title received';
+          node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+          node.error(`${sonosFunction}  - ${msgShort} Details: ` + JSON.stringify(response));
           return;
         }
       }
       // Output data
+      node.debug('got valid song info');
       msg.song = response;
       msg.artist = artist;
-      msg.title = title;
       msg.albumArtURL = albumArtURL;
-
-      if (chain) {
-        node.debug('got current song information');
-        getSonosMediaInfo(node, msg, sonosPlayer, chain);
-      } else {
-        node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
-        node.debug(`ok:${sonosFunction}`);
-        node.send(msg);
+      switch (outputType) {
+        case 'get_songmedia':
+          // part of a chain, dont send message
+          node.debug('continue async');
+          msg.title = title;
+          getMediaInfo(node, msg, sonosPlayer, 'get_songmedia');
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.title = title;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = title;
+          node.send(msg);
       }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  function getSonosMediaInfo (node, msg, sonosPlayer, chain) {
-    //   changes msg.media
-    const sonosFunction = 'media info';
-    let errorShort = 'invalid media info received';
+  /** Gets the media info and either outputs or starts another asyn request (chaining).
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.queue_active, msg.payload or msg.media
+  */
+  function getMediaInfo (node, msg, sonosPlayer, outputType) {
+    const sonosFunction = 'get media info';
+    let msgShort = '';
     sonosPlayer.avTransportService().GetMediaInfo().then(response => {
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid media info received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
       const uri = response.CurrentURI;
@@ -300,49 +487,81 @@ module.exports = function (RED) {
         response.queue_active = false;
       }
 
-      msg.media = response;
-      if (chain) {
-        node.debug('got valid media info');
-        getSonosPositionInfo(node, msg, sonosPlayer, chain);
-      } else {
-        node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
-        node.debug(`ok:${sonosFunction}`);
-        node.send(msg);
+      // output chaining, to seperate msg property or to msg.payload
+      node.debug('got valid media info');
+      switch (outputType) {
+        case 'get_songmedia':
+          // part of a chain, dont send message
+          msg.media = response;
+          node.debug('continue');
+          getPositionInfo(node, msg, sonosPlayer, 'get_songmedia');
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.media = response;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
       }
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }
 
-  function getSonosPositionInfo (node, msg, sonosPlayer, chain) {
-    //   changes msg.position
+  /** Gets the position inf and either outputs or starts another asyn request (chaining).
+  * @param  {Object} node current node
+  * @param  {Object} msg incoming message
+  * @param  {Object} sonosPlayer sonos player object
+  * @param  {String} outputType selector for output type: chainidentifier, 'seperate' and default = 'payload'
+  * changes msg.position or msg.payload
+  */
+  function getPositionInfo (node, msg, sonosPlayer, outputType) {
     const sonosFunction = 'position info';
-    let errorShort = 'invalid position info received';
+    let msgShort = '';
     sonosPlayer.avTransportService().GetPositionInfo().then(response => {
       node.debug(JSON.stringify(response));
       if (response === null || response === undefined) {
-        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-        node.error(`${sonosFunction} - ${errorShort} Details: respose ->` + JSON.stringify(response));
+        msgShort = 'invalid position info received';
+        node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+        node.error(`${sonosFunction} - ${msgShort} Details: respose ->` + JSON.stringify(response));
         return;
       }
+
+      // output chaining, to seperate msg property or to msg.payload
       node.debug('got valid positon info');
-      msg.position = response;
-      // Send output
-      if (chain) {
-        node.debug('got valid position info');
-        node.status({ fill: 'green', shape: 'dot', text: 'ok:get songmedia' });
-        node.debug('ok:get songmedia');
-      } else {
-        node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
-        node.debug(`ok:${sonosFunction}`);
+      switch (outputType) {
+        case 'get_songmedia':
+          // end of chain
+          msg.position = response;
+          node.status({ fill: 'green', shape: 'dot', text: 'ok:get songmedia' });
+          node.debug('ok:get songmedia');
+          node.send(msg);
+          break;
+        case 'separate':
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.position = response;
+          node.send(msg);
+          break;
+        default:
+        // payload
+          node.status({ fill: 'green', shape: 'dot', text: `ok:${sonosFunction}` });
+          node.debug(`ok:${sonosFunction}`);
+          msg.payload = response;
+          node.send(msg);
       }
-      node.send(msg);
     }).catch(err => {
-      errorShort = 'error caught from response';
-      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${errorShort}` });
-      node.error(`${sonosFunction} - ${errorShort} Details: ` + JSON.stringify(err));
+      msgShort = 'error caught from response';
+      node.status({ fill: 'red', shape: 'dot', text: `error:${sonosFunction} - ${msgShort}` });
+      node.error(`${sonosFunction} - ${msgShort} Details: ` + JSON.stringify(err));
     });
   }